refact: extrair UI de expansão de skill para skill-expand-ui.mjs (Fase 3 — Commit 3/4)
